Move document uploading into DocumentService
